Move preparePathForShell into TerminalInstance
